Updates to stripToken
