+++ server/stripToken.js
@@ -1,5 +1,10 @@
+let JWT = require('./JWT')
+
 module.exports = async (req, res, next) => {
-  delete req.query.token
-  delete req.body.token
-  return next()
+  try {
+    JWT.setTokenPayload(req)
+    return next()
+  } catch (e) {
+    res.status(401).send({ message: e.message || e })
+  }
 }
